Add a way to apply rule overrides
They get stored in a separate json file, get loaded at runtime and
replace an existing rule.

In theory rules can be appended by replacing a non-existing rule,
and rules can be removed by replacing it with a dummy one. But
the format can be extended to make this more clear if needed...