hcert: Skip rules not valid at the current time
The official rules contain rules that are no longer valid in case
there is a rule transition. When checking the hcert we need to use only
the valid rules.

When trying to figure out for how long the HCERT is valid we need to use a fixed
set of rules though, since otherwise the the cahnged set of rules could make an
already invalid HCERT valid again after some time, which conflicts with our
assumptions in the binary search.

To avoid this use the initial check time to select the set of rules and when searching
for the date where the HCERT becomes invalid always use the same set of rules.