tests: Add proper testing for the hcert validation
The main issue here is that hcert-kotlin will use the current time to
check if the rules/trustlist etc are valid and those expire after some days.

To work around that use a date mocking library and wrap all hcert-kotlin calls
with some mocking code. This way we can set a fixed time for everything in
the tests.

For this we introduce a trustDate which gets used for validating the cbor files
and for selecting the active set of rules as well.

This also exposed a problem when there are no rules and the code trying to figure
out for hwo long the hcert is valid wont find an end date, resulting in a timestamp
overflow. To fix detect this case and give up on finding an end date.