hcert: also take the hcert expiration into account when computing validUntil
In case the rules say it is valid for, let's say, two years, but the hcert
expires in one month, then the hcert should be the date displayed to the user.

This depends on patch that isn't commited upstream yet, so update hcert-kotlin
from our fork and add some code that can deal with both versions.

Fixes #1