Move validation to our manually managed rules list
Instead of decoding the official cbor/binary list we pull an already
decoded version called "rules.json" from the server. It has the same
format as the CBOR variant to keep it simple.

This is needed since the official business rules have stopped being updated
for some reason and the TU Graz rules also deviate from the entry test rules.

So we manage things ourselves now.

The tests had to be slightly adjusted and the rules validFrom fields as well
to get something meaningfull out of the dummy test certificate. Since we can't
simply create new HCERTs and have to pick the upstream ones we are forced to
mock everything around it..

And validFrom/validUntil fields are removed because they are missing in the JSON
case. This unifies both sources and the data wasn'T really used anyway.