Get rid of the HTTP wrapper interface for HCERT validation
We are likely not going to go back to server validation, so just use the
JS API directly.

This gives us more insight into the validation data.

If the validation throws then it's due to some infrastructure not working.
If it returns a result that is not valid then it's due to the HCERT not being valid.