Allow validating against multiple rule sets
validate() now takes a list of region identifiers and the validation
result contains a separate result object for each reagion.

The main result now contains the person information and if the HCERT itself
plus the signature is valid/not expired.

The region results contain the result of the rule checks and for how long
the HCERT is valid for that particular region (taking into account the rules and
the expiration date of the HCERT itself)

The only current validate() user code is ported to the previous default,
it only passes ET and checks that ET is valid.